+++ fastkml/links.py
@@ -32,6 +32,8 @@ from fastkml.kml_base import _BaseObject
 from fastkml.registry import RegistryItem
 from fastkml.registry import registry
 
+__all__ = ["Icon", "Link"]
+
 
 class Link(_BaseObject):
     """
@@ -124,7 +126,7 @@ class Link(_BaseObject):
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="href",
         node_name="href",
         classes=(str,),
@@ -135,7 +137,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="refresh_mode",
         node_name="refreshMode",
         classes=(RefreshMode,),
@@ -147,7 +149,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="refresh_interval",
         node_name="refreshInterval",
         classes=(float,),
@@ -159,7 +161,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="view_refresh_mode",
         node_name="viewRefreshMode",
         classes=(ViewRefreshMode,),
@@ -171,7 +173,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="view_refresh_time",
         node_name="viewRefreshTime",
         classes=(float,),
@@ -183,7 +185,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="view_bound_scale",
         node_name="viewBoundScale",
         classes=(float,),
@@ -195,7 +197,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="view_format",
         node_name="viewFormat",
         classes=(str,),
@@ -207,7 +209,7 @@ registry.register(
 registry.register(
     Link,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="http_query",
         node_name="httpQuery",
         classes=(str,),