+++ fastkml/styles.py
@@ -62,6 +62,19 @@ from fastkml.registry import registry
 
 logger = logging.getLogger(__name__)
 
+__all__ = [
+    "BalloonStyle",
+    "HotSpot",
+    "IconStyle",
+    "LabelStyle",
+    "LineStyle",
+    "Pair",
+    "PolyStyle",
+    "Style",
+    "StyleMap",
+    "StyleUrl",
+]
+
 
 class StyleUrl(_XMLObject):
     """
@@ -143,7 +156,7 @@ class StyleUrl(_XMLObject):
 registry.register(
     StyleUrl,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="url",
         node_name="styleUrl",
         classes=(str,),
@@ -229,7 +242,7 @@ class _ColorStyle(_BaseObject):
 registry.register(
     _ColorStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="color",
         node_name="color",
         classes=(str,),
@@ -241,7 +254,7 @@ registry.register(
 registry.register(
     _ColorStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="color_mode",
         node_name="colorMode",
         classes=(ColorMode,),
@@ -513,7 +526,7 @@ class IconStyle(_ColorStyle):
 registry.register(
     IconStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="scale",
         node_name="scale",
         classes=(float,),
@@ -525,7 +538,7 @@ registry.register(
 registry.register(
     IconStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="heading",
         node_name="heading",
         classes=(float,),
@@ -537,7 +550,7 @@ registry.register(
 registry.register(
     IconStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="icon",
         node_name="Icon",
         classes=(Icon,),
@@ -548,7 +561,7 @@ registry.register(
 registry.register(
     IconStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="hot_spot",
         node_name="hotSpot",
         classes=(HotSpot,),
@@ -646,7 +659,7 @@ class LineStyle(_ColorStyle):
 registry.register(
     LineStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="width",
         node_name="width",
         classes=(float,),
@@ -755,7 +768,7 @@ class PolyStyle(_ColorStyle):
 registry.register(
     PolyStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="fill",
         node_name="fill",
         classes=(bool,),
@@ -767,7 +780,7 @@ registry.register(
 registry.register(
     PolyStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="outline",
         node_name="outline",
         classes=(bool,),
@@ -869,7 +882,7 @@ class LabelStyle(_ColorStyle):
 registry.register(
     LabelStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="scale",
         node_name="scale",
         classes=(float,),
@@ -1021,7 +1034,7 @@ class BalloonStyle(_BaseObject):
 registry.register(
     BalloonStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="bg_color",
         node_name="bgColor",
         classes=(str,),
@@ -1033,7 +1046,7 @@ registry.register(
 registry.register(
     BalloonStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="text_color",
         node_name="textColor",
         classes=(str,),
@@ -1045,7 +1058,7 @@ registry.register(
 registry.register(
     BalloonStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="text",
         node_name="text",
         classes=(str,),
@@ -1056,7 +1069,7 @@ registry.register(
 registry.register(
     BalloonStyle,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="display_mode",
         node_name="displayMode",
         classes=(DisplayMode,),
@@ -1148,7 +1161,7 @@ class Style(_StyleSelector):
 registry.register(
     Style,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="styles",
         node_name="Style",
         classes=(
@@ -1258,7 +1271,7 @@ class Pair(_BaseObject):
 registry.register(
     Pair,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="key",
         node_name="key",
         classes=(PairKey,),
@@ -1269,7 +1282,7 @@ registry.register(
 registry.register(
     Pair,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="style",
         node_name="Style",
         classes=(
@@ -1391,7 +1404,7 @@ class StyleMap(_StyleSelector):
 registry.register(
     StyleMap,
     RegistryItem(
-        ns_ids=("kml",),
+        ns_ids=("kml", ""),
         attr_name="pairs",
         node_name="Pair",
         classes=(Pair,),
@@ -1399,15 +1412,3 @@ registry.register(
         set_element=xml_subelement_list,
     ),
 )
-
-
-__all__ = [
-    "BalloonStyle",
-    "IconStyle",
-    "LabelStyle",
-    "LineStyle",
-    "PolyStyle",
-    "Style",
-    "StyleMap",
-    "StyleUrl",
-]